HW06 corr2. JdbcUserRepository return null if updated user does not exist.

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -47,11 +47,12 @@ public class JdbcUserRepository implements UserRepository {
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             user.setId(newKey.intValue());
-        } else  {
-//            TODO : if update user which is not exist return null
-            namedParameterJdbcTemplate.update(
+        } else {
+            if (namedParameterJdbcTemplate.update(
                     "UPDATE users SET name=:name, email=:email, password=:password, " +
-                            "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource);
+                            "registered=:registered, enabled=:enabled, calories_per_day=:caloriesPerDay WHERE id=:id", parameterSource) == 0) {
+                return null;
+            }
             jdbcTemplate.update("DELETE FROM user_roles WHERE user_id=?", user.getId());
         }
         updateRoles(user);
@@ -92,9 +93,9 @@ public class JdbcUserRepository implements UserRepository {
 
     private void updateRoles(User user) {
         Set<Role> roleSet = user.getRoles();
-        if(!CollectionUtils.isEmpty(roleSet)) {
+        if (!CollectionUtils.isEmpty(roleSet)) {
             jdbcTemplate.batchUpdate("INSERT INTO user_roles (user_id, role) VALUES (?, ?)", roleSet, roleSet.size(),
-                    (ps,role) -> {
+                    (ps, role) -> {
                         ps.setInt(1, user.getId());
                         ps.setString(2, role.name());
                     });
